refactor: SM-172 컴포넌트 디렉토리 import export 모듈 리팩토링
