+++ src/components/base/PageContainerWithBackground/index.tsx
@@ -1,1 +1,1 @@
-export { StyledPageContainerWithBackground as Default } from './styled';
+export { StyledPageContainerWithBackground as default } from './styled';
