+++ src/components/base/index.ts
@@ -1,8 +1,12 @@
 export { default as Button } from '@base/Button';
-export { default as Text } from '@base/Text';
 export { default as Divider } from '@base/Divider';
 export { default as Icon } from '@base/Icon';
 export { default as Image } from '@base/Image';
 export { default as Input } from '@base/Input';
 export { default as Modal } from '@base/Modal';
+export { default as PageContainerWithBackground } from '@base/PageContainerWithBackground';
 export { default as SectionDivider } from '@base/SectionDivider';
+export { default as Select } from '@base/Select';
+export { default as Text } from '@base/Text';
+export { default as Tooltip } from '@base/Tooltip';
+export { default as Uploader } from '@base/Uploader';
