+++ src/components/compound/Carousel/CarouselItem/index.tsx
@@ -1,6 +1,5 @@
-import Image from '@base/Image';
-import Text from '@base/Text';
-import IconButton from '@compound/IconButton';
+import { Image, Text } from '@base';
+import { IconButton } from '@compound';
 import type { ReactElement } from 'react';
 import { useCallback } from 'react';
 import { StyledContainer, StyledWrapper } from './styled';
