--- src/components/compound/ImageButton/styled.ts
@@ -1,1 +1,0 @@
-export {};
