+++ src/components/compound/index.tsx
@@ -1,1 +1,10 @@
-export {};
+export { default as Album } from '@compound/Album';
+export { default as Carousel } from '@compound/Carousel';
+export { default as IconButton } from '@compound/IconButton';
+export { default as IconToggle } from '@compound/IconToggle';
+export { default as ImageButton } from '@compound/ImageButton';
+export { default as Loader } from '@compound/Loader';
+export { default as MenuTab } from '@compound/MenuTab';
+export { default as TextButton } from '@compound/TextButton';
+export { default as TextToggle } from '@compound/TextToggle';
+export { default as UserInput } from '@compound/UserInput';
