+++ src/components/domain/CocktailList/index.tsx
@@ -3,7 +3,7 @@ import type { ReactElement } from 'react';
 import { Text } from '@base';
 import { StyledContainer, StyledMotionWrapper } from './styled';
 import type { CocktailListProps } from './types';
-import Album from '@compound/Album';
+import { Album } from '@compound';
 
 const DEFAULT_MSG = '죄송합니다. 추천 칵테일이 존재하지 않습니다.';
