+++ src/components/domain/IngredientSelectModal/index.tsx
@@ -1,7 +1,7 @@
 import type { ReactElement } from 'react';
 import { useState, useEffect, useCallback, Children } from 'react';
 import { Modal, Text } from '@base';
-import MenuTab from '@compound/MenuTab';
+import { MenuTab } from '@compound';
 import IngredientToggleList from '@domain/IngredientToggleList';
 import type { IngredientSelectModalProps } from './types';
 import {
