--- src/components/domain/RatingStar/style.ts
@@ -1,1 +1,0 @@
-export {};
